# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Tuneverse Writing API",
    "description": "API for Tuneverse Writing"
  },
...
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/api/user/cheques": {
      "get": {
        "tags": ["Cheque"],
        "summary": "Get all cheques",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Cheque"],
        "summary": "Create cheque",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Cheque object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "payment": {
                  "type": "string"
                },
                "accountNumber": {
                  "type": "string"
                },
                "userId": {
                  "type": "integer"
                },
                "chequeTypeId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created cheque"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/cheques/{id}": {
      "get": {
        "tags": ["Cheque"],
        "summary": "Get cheque by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/cheques/{id}": {
      "put": {
        "tags": ["Cheque"],
        "summary": "Change cheque by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "countryId": {
                  "type": "integer"
                },
                "userRoleId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/cheques/{id}": {
      "delete": {
        "tags": ["Cheque"],
        "summary": "Delete cheque by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/cheques/{id}/logicalDelete": {
      "patch": {
        "tags": ["Cheque"],
        "summary": "Logical delete cheque by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/cheques/{id}/logicalRestore": {
      "patch": {
        "tags": ["Cheque"],
        "summary": "Logical restore cheque by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/chequeTypes": {
      "get": {
        "tags": ["ChequeType"],
        "summary": "Get all chequeTypes",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/chequeTypes": {
      "post": {
        "tags": ["ChequeType"],
        "summary": "Create chequeType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "ChequeType object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created chequeType"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/chequeTypes/{id}": {
      "get": {
        "tags": ["ChequeType"],
        "summary": "Get chequeType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ChequeType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/chequeTypes/{id}": {
      "put": {
        "tags": ["ChequeType"],
        "summary": "Change chequeType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ChequeType ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/chequeTypes/{id}": {
      "delete": {
        "tags": ["ChequeType"],
        "summary": "Delete chequeType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ChequeType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/chequeTypes/{id}/logicalDelete": {
      "patch": {
        "tags": ["ChequeType"],
        "summary": "Logical delete chequeType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ChequeTypes ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/chequeTypes/{id}/logicalRestore": {
      "patch": {
        "tags": ["ChequeType"],
        "summary": "Logical restore chequeType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cheque ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/countries": {
      "get": {
        "tags": ["Country"],
        "summary": "Get all countries",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/countries": {
      "post": {
        "tags": ["Country"],
        "summary": "Create country",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Country object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "countryCode": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created country"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/countries/{id}": {
      "get": {
        "tags": ["Country"],
        "summary": "Get country by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/countries/{id}": {
      "put": {
        "tags": ["Country"],
        "summary": "Change country by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "countryCode": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/countries/{id}/logicalDelete": {
      "patch": {
        "tags": ["Country"],
        "summary": "Logical delete country by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/countries/{id}/logicalRestore": {
      "patch": {
        "tags": ["Country"],
        "summary": "Logical restore country by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}